+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Changed the hint path for filter, controller and service sub-generators
 - Created a new utils config module to handle stuff about the callback config
 - The view sub-generator will now automatically update the main less file for import
+- The config sub-task will now update the package file as well to store the callback config
 
 ## [0.7.0]
 ### Added